+++ changelog.md
@@ -4,6 +4,7 @@
 - update: x265 3.1+15-a092e82 Wolfberry
 - update: MP4Box 0.8.0-rev69-5fe3ec1 Wolfberry
 - update: ffmpeg 4.2.1 Wolfberry
+- update: L-Smash Works 2019-09-14 HolyWu
 - the filter list in the main dialog shows shorter filter names,
   for example 'BicubicResize' instead of 'Resize BicubicResize'
 - SubtitleEdit is now included